Refactor path of joined missions component
