+++ src/components/my-profile/MyProfile.jsx
@@ -1,6 +1,6 @@
 import './myProfile.css';
 import React from 'react';
-import JoinedMissions from './JoinedMissions';
+import JoinedMissions from './JoinedMissions/JoinedMissions';
 import ReservedRockets from './ReservedRockets/ReservedRockets';
 
 const MyProfile = () => (
